+++ tests/test_kernel32/test_handle.py
@@ -1,6 +1,12 @@
+import os
+import tempfile
+from errno import EBADF
+
 from pywincffi.core import dist
 from pywincffi.dev.testutil import TestCase
-from pywincffi.kernel32.handle import GetStdHandle
+from pywincffi.exceptions import InputError
+from pywincffi.kernel32.handle import (
+    GetStdHandle, CloseHandle, handle_from_file)
 
 
 class TestGetStdHandle(TestCase):
@@ -24,3 +30,38 @@ class TestGetStdHandle(TestCase):
             GetStdHandle(library.STD_ERROR_HANDLE),
             library.GetStdHandle(library.STD_ERROR_HANDLE)
         )
+
+
+class TestGetHandleFromFile(TestCase):
+    def test_fails_if_not_a_file(self):
+        with self.assertRaises(InputError):
+            handle_from_file(0)
+
+    def test_fails_if_file_is_not_open(self):
+        fd, _ = tempfile.mkstemp()
+        test_file = os.fdopen(fd, "r")
+        test_file.close()
+
+        with self.assertRaises(InputError):
+            handle_from_file(test_file)
+
+    def test_opens_correct_file_handle(self):
+        fd, path = tempfile.mkstemp()
+        os.close(fd)
+
+        test_file = open(path, "w")
+        handle = handle_from_file(test_file)
+
+        CloseHandle(handle)
+
+        # If CloseHandle() was passed the same handle
+        # that test_file is trying to write to the file
+        # and/or flushing it should fail.
+        try:
+            test_file.write("foo")
+            test_file.flush()
+        except (OSError, IOError, WindowsError) as error:
+            # EBADF == Bad file descriptor (because CloseHandle closed it)
+            self.assertEqual(error.errno, EBADF)
+        else:
+            self.fail("Expected os.close(%r) to fail" % fd)
