+++ tests/test_kernel32/test_io.py
@@ -1,13 +1,8 @@
-import os
-import tempfile
-from errno import EBADF
-
 from pywincffi.dev.testutil import TestCase
-from pywincffi.exceptions import WindowsAPIError, InputError
+from pywincffi.exceptions import WindowsAPIError
 from pywincffi.kernel32.handle import CloseHandle
 from pywincffi.kernel32.io import (
-    CreatePipe, WriteFile, ReadFile,
-    PeekNamedPipe, PeekNamedPipeResult, handle_from_file)
+    CreatePipe, WriteFile, ReadFile, PeekNamedPipe, PeekNamedPipeResult)
 
 # For pylint on non-windows platforms
 try:
@@ -144,38 +139,3 @@ class TestPeekNamedPipe(PipeBaseTestCase):
         result = PeekNamedPipe(reader, 0)
         self.assertEqual(
             result.lpTotalBytesAvail, bytes_written - (read_bytes * 2))
-
-
-class TestGetHandleFromFile(TestCase):
-    def test_fails_if_not_a_file(self):
-        with self.assertRaises(InputError):
-            handle_from_file(0)
-
-    def test_fails_if_file_is_not_open(self):
-        fd, _ = tempfile.mkstemp()
-        test_file = os.fdopen(fd, "r")
-        test_file.close()
-
-        with self.assertRaises(InputError):
-            handle_from_file(test_file)
-
-    def test_opens_correct_file_handle(self):
-        fd, path = tempfile.mkstemp()
-        os.close(fd)
-
-        test_file = open(path, "w")
-        handle = handle_from_file(test_file)
-
-        CloseHandle(handle)
-
-        # If CloseHandle() was passed the same handle
-        # that test_file is trying to write to the file
-        # and/or flushing it should fail.
-        try:
-            test_file.write("foo")
-            test_file.flush()
-        except (OSError, IOError, WindowsError) as error:
-            # EBADF == Bad file descriptor (because CloseHandle closed it)
-            self.assertEqual(error.errno, EBADF)
-        else:
-            self.fail("Expected os.close(%r) to fail" % fd)
